tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/changing_animation_props.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - changing animation-duration during animation</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="The values used for the animation properties are
   snapshotted at the time the animation starts">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            margin-left: 0px;
         }
         to {
            margin-left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation: slide 1s;
      }
      .longer {
         animation-duration: 2s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxFactor = 0.01;
            return (Math.abs(observed - actual) < maxFactor * actual);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            div.className = 'test longer';
         });
         div.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            var passed = aboutEqual(evt.elapsedTime, 1);
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Below there should be a blue square sliding right for 1 second and then
   jumping back to its initial position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
